Do not assert that the ruleset does not contain unreachable rules
